+++ src/utils/commons.js
@@ -1,22 +1,10 @@
-import dayjs from 'dayjs';
-import duration from 'dayjs/plugin/duration';
-
-dayjs.extend(duration);
-
-const humanizeDate = (date, dateformat) => date ? dayjs(date).format(dateformat) : '';
-
-const calculateDuration = (start, stop) => {
-  const eventTime = dayjs.duration(dayjs(stop) - dayjs(start), 'millisecond');
-
-  if (eventTime.$d.days) {
-    return eventTime.format('DD[d] HH[h] mm[M]');
-  } else if (!eventTime.$d.days && eventTime.$d.hours) {
-    return eventTime.format('HH[h] mm[M]');
-  }
-  return eventTime.format('mm[M]');
-
-};
-
+//@ts-check
+/**
+ *
+ * @param {number} [min = 1] minimum value
+ * @param {number} [max = 100] maximum value
+ * @returns {number} random number
+ */
 const getRandomInteger = (min = 1, max = 100) => {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -25,8 +13,19 @@ const getRandomInteger = (min = 1, max = 100) => {
   return Math.floor(result);
 };
 
+/**
+ *
+ * @param {Array} elements
+ * @returns {any} random element
+ */
 const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
 
+/**
+ * @generator generete unique numbers
+ * @param {number} [min = 1] min minimum value
+ * @param {number} [max = 100] max maximum value
+ * @yields {number} unique number
+ */
 const getRandomIntGenerator = (min = 1, max = 100) => {
   const usedIds = new Set();
 
@@ -41,19 +40,21 @@ const getRandomIntGenerator = (min = 1, max = 100) => {
   };
 };
 
-const checkEcsKeydownPress = (event, func, handkerFunc) => {
+/**
+ *
+ * @param {KeyboardEvent} event
+ * @param {function} func
+ */
+const checkEcsKeydownPress = (event, func) => {
   if (event.key === 'Escape') {
     event.preventDefault();
     func();
-    document.removeEventListener('keydown', handkerFunc);
   }
 };
 
 export {
-  calculateDuration,
   checkEcsKeydownPress,
   getRandomArrayElement,
   getRandomInteger,
   getRandomIntGenerator,
-  humanizeDate,
 };
